+++ packages/three/index.js
@@ -1,2 +1,3 @@
 export * from "./src/ThreeRender";
+export * from "./src/Instances";