+++ packages/three/src/ThreeRender.js
@@ -41,8 +41,6 @@ export const PM_ThreeVisible = superclass => class extends superclass {
 
 }
 
-
-
 //------------------------------------------------------------------------------------------
 //-- ThreeCamera  --------------------------------------------------------------------------
 //------------------------------------------------------------------------------------------
@@ -50,7 +48,6 @@ export const PM_ThreeVisible = superclass => class extends superclass {
 export const PM_ThreeCamera = superclass => class extends superclass {
     constructor(...args) {
         super(...args);
-        console.log("ThreeCamera");
 
         this.cameraTranslation = [0,0,0]; // position of the camera relative to the pawn
         this.cameraRotation = q_identity();