+++ packages/three/CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+-- ThreeInstanceManager
+-- PM_ThreeInstanced
+
+### Removed
+-- Layers and Raycasting
+-- Post-processing
+-- BVH
+
 
 ## [1.2.0] - 2022-4-20
 ### Changed